Fix and unify CreateOrUpdateXXX functions for karmadactl
